% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode.r
\name{FOS}
\alias{FOS}
\title{The classic FOS method}
\usage{
FOS(x, alpha = 0.0027)
}
\arguments{
\item{x}{Input data vector}

\item{alpha}{False alarm rate}
}
\value{
the lower and upper confidence bounds.
\itemize{
 \item{ limits:}{ The lower and upper confidence limit}
 \item{ pn :}{ The suggested nominal coverage probability based on the sample size}
 }
}
\description{
The classic FOS method
}
\examples{
FOS(x1, alpha = 0.0027)
}
